requirements_local.txt
backcall==0.1.0
certifi==2018.4.16
chardet==3.0.4
decorator==4.3.0
dj-database-url==0.5.0
Django==2.1
django-cors-headers==2.4.0
django-filter==2.0.0
djangorestframework==3.8.2
gunicorn==19.9.0
idna==2.7
ipdb==0.11
ipython==6.5.0
ipython-genutils==0.2.0
jedi==0.12.1
Markdown==2.6.11
oauthlib==2.1.0
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.5
ptyprocess==0.6.0
py-trello==0.11.3
Pygments==2.2.0
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.23
wcwidth==0.1.7
whitenoise==3.3.0
